host/requirements.txt
numpy>=1.24
pyserial>=3.5
matplotlib>=3.7
